Change canceButtonIndex and destructiveButtonIndex to NSInteger to match UIActionSheet

+++ IBActionSheetSample/IBActionSheetSample/IBActionSheet.h
@@ -112,7 +112,7 @@ typedef void (^IBActionCallback)(IBActionSheet *actionSheet, NSInteger buttonInd
 @property (weak) id <IBActionSheetDelegate> delegate;
 @property IBActionSheetButtonResponse buttonResponse;
 @property BOOL visible, hasCancelButton, hasDestructiveButton, shouldCancelOnTouch, blurBackground; // Background Blur is only currently availble on iOS 8.
-@property int cancelButtonIndex, destructiveButtonIndex;
+@property NSInteger cancelButtonIndex, destructiveButtonIndex;
 @property (nonatomic, copy) IBActionCallback callback;
 
 @end
